Maintain percent encoding in user credentials
When using credentials via URL, any encoded characters need to remain
encoded. An example of this would be a user name that has an @ symbol.
The @ symbol must be encoded as %40. Before this PR, uri methods related
to user credentials would decode %40 back into @. This PR changes
getAuthority and getUserInfo calls to getRawAuthority and getRawUserInfo,
which preserve percent encoded characters.