Enable the use of multiple name elements
Not every sequence entry in the WMO xml table D files are processed,
only those that contain a specific child element (ElementName_E). For
older WMO tables, sometimes that element isn't used, and so the sequence
is skipped when reading in the table. However, there are cases where we
should probably be looking for another element to use in addition to the
one we currently look for. An example of this is ExistingElementName_E,
and we do see entries where ElementName_E is not used, but
ExistingElementName_E is. Skipping these entries has caused errors, so
I'm pretty sure we need to keep those. There are potential sequence
entries we skip which do not use either one of those element names, but
I'm not sure what to do with them quite yet. They might be cruft
describing table headings from actual tables used in the word and pdf
documents, but they have FXY1 and FXY2, so maybe they should not be
skipped.

As an example, with this change we go from using 4886 out of 5321
potential sequences in WMO Table D v16 to using 5294.