Fix addition of javax.servlet-api dependency
After bumping to mainline gretty and having our client tests use a
gretty managed local server for testing, we picked up (for some reason)
a dependency on javax.servlet-api in the poms of the client artifacts
(httpservices, for example). This causes a problem when building the TDS,
for example, as we do not want a javax.servlet-api compile dependency
being added directly to the war. These changes remove the unwanted
dependency. This might be a bug in gradle, but we're running 3.5, so
maybe once we upgrade to 6, things will work without a special
workaround.