Close Datasets that do not have the StationFeature sought
Make sure that datasets that do not contain a given StationFeature
are release/closed, otherwise, if one or more caches are enabled (i.e.
NetcdfDataset's NetcdfFile Cache or RandomAccessFile's cache), the
file handle will be locked and new file handles will be opened each
pass through the collection. These exist in the cache until the OS runs
out of file handles, or until the cache is forcefully purged, and we
all have a bad time when that happens.

Fixes Unidata/netcdf-java#320