Ensure projection constructCopy uses base values
In several projection classes in ucar.unidata.geoloc.projection, the
constructCopy methods were using computed values to pass back into the
main constructor. Typically, this was due to a change in units from
degrees (as input into the constructor), to radians (as needed by the
class), then back to degrees (as needed by constructCopy). The result
was that the equals methods were failing due to floating point math.
This commit fixes the issue by ensureing constructCopy uses the values
as they were passed in through the constructor.