Manage size of dap GET request prefetch URI
When prefetching data for a dap request, we need to manage the size of
the URI to avoid a "414 URI Too Long" response from the server. Tricky
because the maximum size of a GET request URI isn't part of the RFC
2616, and depends on the server side webserver stack. 4096 bytes appears
to be a good size based on apache, nginx, iis, tomcat defaults.

Fixes Unidata/netcdf-java#413