Extend CdmS3Uri to include the concept of a delimiter
A delimiter can be used to indicate that there is a hierarchical
structure in the keys of a bucket. For example, if files are uploaded to
an object store and are given a key name equal to their path on disk,
the delimiter would likely be "/" or "\". The extension is made using
the addition of a URI fragment, something like `#delimiter="/"`