Allow a DtCoverageDataset name to be set within NcML
When opening a DtCoverageDataset using the string representation of NcML
(i.e. not file-based), the DtCoverageDataset name is null (as the
underlying NetcdfDataset object location is null). This is a problem in
the for the TDS when NcML is defined in a configuration catalog, as we
end up throwing a NPE. This change enables the NcML ID to be used as the
DtCoverageDataset name in cases where there is a null location.