Don't delete .ar2v files in nexrad hires test.
Fixes a new Jenkins failure due to test worker upgrade, which changed
the order some of the tests were run. Because this test deleted a test
file, and the test that needed that file began running after this test,
the test that needed the file began to fail with a file not found
exception.