Fix failing S3 inventory test
Fix failing S3 inventory test, but only run those types of tests on
jenkins and not against PRs. The reasons is that new objects get added
to the top level of the object store buckets, and end up breaking the
test. I think the tests are still useful, as they check that the
delimiter code still works, but we don't want to hold up pull requests
because a new file was added to the top level of a bucket that we do not
control.