Adjust AWS top level bucket object count test
The top level of the AWS bucket used in TestControllerS3 does not have a
fixed number of objects. It seems sometimes a file is added or removed
without warning. The test that counts top level objects has been updated
to check the count as an "at most" test against an expected number instead
of an exact count.