Fix root publishing script
When making changes to publishing of the fatJar artifacts to the
download raw repo instead of the maven repo (Unidata/netcdf-java#777),
I removed some code that was necessary for passing credentials to the
code that handles publishing to the maven repo. This PR restores that
code and now things work again.