Make sure that DAP4 xml strings do not contain NUL characters.
The current remotetest/d4ts servlet is failing because it is
including the trailing NUL characters in the DMR XML metadata.
This is illegal XML.  Fix, of course, is to suppress NUL
characters.

This change has been tested on remotetest, but it is only indirectly
tested in the existing DAP4 tests.

This is the first in a probably long list of DAP4 fixes.