GEOSTranform better test if past limb
As previously coded, GEOSTransform was providing non-null results for locations that were just past the visible limb of the planet but not more than 90° away from the below-satellite point.

I have added a better test to determine if such lon-lat points should be ignored based on the "GOES R SERIES PRODUCT DEFINITION AND USERS’ GUIDE (PUG)", sects. 5.1.2.8.1-5.1.2.8.2 (pp. 21-24) which can be dowloaded at https://www.goes-r.gov/resources/docs.html

I have tested this for the GOES orientation but not the GEOS, as I do not have sample datasets for the latter.